Redesign how post replacement works
Always reuse the first entry object created and simply update its values
rather than replacing it with a new object when a post is edited. This
makes it so that object references stay valid, and simplifies a lot of
the replacement logic by eliminating the need for things like
reparenting children.